requirements.txt
dbt-duckdb
pandas

